+++ apps/web/src/app/live/[organizationHandle]/controller.tsx
@@ -1,10 +1,19 @@
-"use client"
+'use client'
 
-import type { CampaignType } from "@/models/Campaign"
-import type { OrganizationType } from "@/models/Organization"
-import { createInitialisedObjectContext } from "@repo/utilities/client"
-import { useRouter, useSearchParams } from "next/navigation"
-import { useEffect } from "react"
+import type { HistoryType } from '@/models/History'
+import { useRouter, useSearchParams } from 'next/navigation'
+import {
+  createContext,
+  type ReactNode,
+  useCallback,
+  useContext,
+  useEffect,
+} from 'react'
+import { useDebouncedCallback } from 'use-debounce'
+import { saveHistory, spin } from "./serverActions"
+import type { CampaignType } from '@/models/Campaign'
+import type { OrganizationType } from '@/models/Organization'
+import { createInitialisedObjectContext } from '@repo/utilities/client'
 
 
 export type CustomerViewState = {
@@ -23,25 +32,131 @@ export type CustomerViewState = {
     rotating: boolean
     current: 'disabled' | 'ready' | 'spinning' | 'finished'
     prizeIndex?: number
-  },
-  historyId: string | null,
+  }
+  historyId: string | null
+  links: 'qr' | 'button'
 }
-
-const [
-  hook,
-  provider,
+export const [
+  useCustomerViewState, CustomerViewStateProvider,
 ] = createInitialisedObjectContext<CustomerViewState>()
 
-export const useCustomerViewState = hook
-export const CustomerViewStateProvider = provider
+type SpinCallbacks = {
+  onStartSpin: () => Promise<void>
+  onEndSpin: () => Promise<void>
+  pushHistoryDebounced: (historyData: Partial<HistoryType>) => void
+  pushHistory: (historyData: Partial<HistoryType>) => Promise<Partial<HistoryType>>
+}
+
+const SpinContext = createContext<SpinCallbacks | undefined>(undefined)
+export function SpinProvider({ children }: { 
+  children: ReactNode 
+}) {
+  const [state, setState] = useCustomerViewState()
+  const router = useRouter()
+
+  const onStartSpin = useCallback(async () => {
+    // Do nothing if we can't spin yet
+    if (state.wheel.current != 'ready') return     
+
+    // Choose a random prize
+    const selected = state.campaigns.selected
+    const { index } = await spin({
+      historyId: state.historyId as string,
+      campaign: state.campaigns.list[selected],
+    })
+    setState({ wheel: { 
+      ...state.wheel, 
+      prizeIndex: index, 
+      current: 'spinning' 
+    }})
+  }, [state])
+
+  const onEndSpin = useCallback(async () => {
+    if (state.wheel.current != 'spinning') return
+    router.push(`/live/${state.organization.handle}/prize?selectedCampaign=${state.campaigns.selected}`)
+    setState({ wheel: { ...state.wheel , current: 'finished' } })
+  }, [state])
+
+  const pushHistory = useCallback(async (historyData: Partial<HistoryType>) => {
+    // Make sure we have an initial historyId
+    if (state.historyId == null) {
+      const newHistoryData = await saveHistory({
+        ...historyData,
+        campaignId: state.campaigns.list[state.campaigns.selected].id,
+        organizationId: state.organization.id,
+      })
+      setState({ historyId: newHistoryData.id })
+      return newHistoryData
+    }
+
+    // Update the history record
+    const newHistoryData = await saveHistory({
+      ...historyData,
+      id: state.historyId,
+    })
+    return newHistoryData
+  }, [state])
+
+  const pushHistoryDebounced = useDebouncedCallback(pushHistory, 800)
+
+  return (
+    // eslint-disable-next-line @typescript-eslint/no-misused-promises
+    <SpinContext.Provider value={{ onStartSpin, onEndSpin, pushHistoryDebounced, pushHistory }}>
+      {children}
+    </SpinContext.Provider>
+  )
+}
+
+export function useSpinCallbacks() {
+  const context = useContext(SpinContext)
+  if (!context) {
+    throw new Error('useSpinCallbacks must be used within a SpinProvider')
+  }
+  return context
+}
+
+export function useGotoRoute () {
+  const router = useRouter()
+  const [state] = useCustomerViewState()
+
+  const resolve = useCallback(function(path: string, options: {
+    selectedCampaign?: number,
+    links?: 'qr' | 'button',
+  }) {
+    const selectedCampaign =  Number(options.selectedCampaign ?? state.campaigns.selected ?? 0)
+    const links = options.links ?? state.links ?? 'qr'
+    const route = `${path}?selectedCampaign=${selectedCampaign}&links=${links}`
+    return route
+  }, [state.campaigns.selected])
+
+  const goto = useCallback(function (path: string, options: {
+    selectedCampaign?: number,
+    links?: 'qr' | 'button',
+  }={}) {
+    const fullRoute = resolve(path, options)
+    router.push(fullRoute)
+  }, [router, state.campaigns.selected])
+
+  return {
+    goto,
+    resolve
+  }
+}
 
 export function useQueryParamUpdateEffect() {
   const [state, setState] = useCustomerViewState()
   const searchParams = useSearchParams()
   useEffect(() => {
     const selected = Number(searchParams.get('selectedCampaign') || 0)
-    if (selected !== state.campaigns.selected) {
-      setState({ campaigns: { ...state.campaigns, selected } })
+    const links = (searchParams.get('links') || 'qr') as 'qr' | 'button'
+    if (selected !== state.campaigns.selected || links !== state.links) {
+      setState({
+        campaigns: {
+          ...state.campaigns,
+          selected,
+        },
+        links,
+      })
     }
   }, [state.campaigns, searchParams])
 }
@@ -49,10 +164,11 @@ export function useQueryParamUpdateEffect() {
 export function useEnforceDefinedHistory() {
   // If we don't have a historyId navigate home
   const [state] = useCustomerViewState()
-  const router = useRouter()
+  const { goto } = useGotoRoute()
   useEffect(() => {
+    // If we don't have a historyId navigate home
     if (state.historyId == null) {
-      router.replace(`/live/${state.organization.handle}/campaigns`)
+      goto(`/live/${state.organization.handle}/campaigns`)
     }
   }, [])
 }