+++ apps/web/src/app/live/[organizationHandle]/layout.tsx
@@ -1,35 +1,25 @@
-import { clamp } from '@repo/utilities'
+
 import type { ReactNode } from 'react'
-import { CustomerViewStateProvider } from './controller'
+import { CustomerViewStateProvider, SpinProvider } from "./controller"
+import { WheelDisplay } from "./WheelDisplay"
+import { InformationDisplay } from './InformationDisplay'
 import { resolveOrganization } from './serverActions'
 
 
 
-export default async function Layout({ 
+export default async function Layout({
   children,
   params,
-  searchParams
-}: { 
+}: {
   children: ReactNode
   params: {
     organizationHandle: string
   }
-  searchParams?: {
-    selected?: string
-  }
 }) {
-  const selectedIndex = Number(searchParams?.selected ?? 0)
   const { organizationHandle } = await params
   const organization = await resolveOrganization(organizationHandle)
-  const selected = clamp(selectedIndex, 0, organization.campaigns.length - 1)
-
   return (
     <div className="w-[100svw] h-[100svh] relative flex bg-slate-200">
-      {/* TODO: Find a good background video for the wheel */}
-      {/* <video autoPlay muted loop playsInline
-        className="bg-slate-200 absolute -z-10 top-1/2 left-1/2 w-full h-full object-cover -translate-x-1/2 -translate-y-1/2"
-      > <source src="/videos/tunnel.mp4" type="video/mp4" />
-      </video> */}
       <CustomerViewStateProvider initial={{
         organization: {
           id: organization.organizationId,
@@ -38,7 +28,7 @@ export default async function Layout({
         },
         campaigns: {
           list: organization.campaigns,
-          selected,
+          selected: 0,
         },
         wheel: {
           animating: false,
@@ -47,8 +37,19 @@ export default async function Layout({
           current: 'disabled',
         },
         historyId: null,
-      }}> 
-        {children}
+        links: 'qr',
+      }}>
+        <SpinProvider>
+          {/* TODO: Find a good background video for the wheel */}
+          {/* <video autoPlay muted loop playsInline
+            className="bg-slate-200 absolute -z-10 top-1/2 left-1/2 w-full h-full object-cover -translate-x-1/2 -translate-y-1/2"
+          > <source src="/videos/tunnel.mp4" type="video/mp4" />
+          </video> */}
+          <WheelDisplay />
+          <InformationDisplay>
+            {children}
+          </InformationDisplay>
+        </SpinProvider>
       </CustomerViewStateProvider>
     </div>
   )