+++ apps/web/src/models/Action.ts
@@ -1,43 +1,36 @@
 
 export const availableActions = [{ 
     label: 'Follow us on Instagram', 
-    value: 'instagram-follow', 
     platform: 'instagram', 
     icon: 'instagram',
     instruction: 'Scan the code below to leave a review'
   }, { 
     label: 'Tag us on Instagram', 
-    value: 'instagram-review', 
     platform: 'instagram', 
     icon: 'instagram',
     instruction: 'Make a new post and include the following Instagram handle',
   }, { 
     label: 'Leave a Google Review', 
-    value: 'google-review', 
     platform: 'google', 
     icon: 'google',
     instruction: 'Scan the code below to leave a review'
   }, { 
     label: 'Leave a Facebook Review', 
-    value: 'facebook-review', 
     platform: 'facebook', 
     icon: 'facebook',
     instruction: 'Scan the code below to leave a review',
   }, { 
     label: 'Follow us on Facebook', 
-    value: 'facebook-follow', 
     platform: 'facebook', 
     icon: 'facebook',
     instruction: 'Scan the code below and follow us',
   }, { 
     label: 'Follow us on TikTok', 
-    value: 'tik-tok-follow', 
     platform: 'tiktok',
     icon: 'tiktok',
     instruction: 'Scan the code below and follow us',
   }, { 
     label: 'Tag us on TikTok', 
-    value: 'tik-tok-tag', 
     platform: 'tiktok',
     icon: 'tiktok',
     instruction: 'Make a new post and include the following TikTok handle',